requirements.txt
black==24.10.0
blinker==1.8.2
click==8.1.7
Flask==3.0.3
gunicorn==23.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
mypy-extensions==1.0.0
packaging==24.1
pathspec==0.12.1
platformdirs==4.3.6
psycopg2-binary==2.9.10
Werkzeug==3.1.2
psycopg[c]; sys_platform == "linux"
psycopg[binary]; sys_platform != "linux"
